+++ react/components/commonEvents.ts
@@ -1,0 +1,8 @@
+// Common OrderPlaced function
+export const orderPlaced = (order: any) => {
+  ga('send', 'event', {
+    eventAction: 'view',
+    eventCategory: 'orderPlaced',
+    ...order,
+  })
+}