+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Add `send` event to each enhanced commerce event.
+
 ## [0.2.0] - 2019-05-08
 
 ### Added